+++ znet/server.go
@@ -134,7 +134,7 @@ func (s *Server) Serve() {
 }
 
 //路由功能：给当前服务注册一个路由业务方法，供客户端链接处理使用
-func (s *Server) AddRouter(msgId uint32, router ziface.IRouter) {
+func (s *Server) AddRouter(msgId uint16, router ziface.IRouter) {
 	s.msgHandler.AddRouter(msgId, router)
 }
